+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Chore
 
 * Update compatibility to include Typesense v29.0
+* Added `NaturalLanguage` operation
+* Added `ExTypesense.flush_analytics/1`
+* Added `ExTypesense.get_analytics_status/1`
+* Added `ExTypesense.get_analytics_events/1`
+
+### Removed
+
+* `Operations.config/2`. Use `Operations.toggle_slow_request_log/2`
 
 ## 2.0.0 (2025.06.01)