+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## major.minor.patch (yyyy.mm.dd)
 
-## 2.1.0 (2026.04.12)
+## 2.1.0 (2026.04.13)
 
 ### Chore